run linter

+++ contracts/echidna/Echidna.sol
@@ -10,8 +10,9 @@ import "./EchidnaUtils.sol";
 //rm -rf crytic-export artifacts &&  npx hardhat clean && npx hardhat compile && echidna . --contract Echidna --config ./contracts/echidna/echidna.conf.yaml
 contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
     address payable immutable _buyer = payable(address(this));
-    address payable immutable _seller = payable(0x2B5AD5c4795c026514f8317c7a215E218DcCD6cF); // used to sign with pk 2
-    
+    address payable immutable _seller =
+        payable(0x2B5AD5c4795c026514f8317c7a215E218DcCD6cF); // used to sign with pk 2
+
     address _conduit;
     bytes32 _conduitKeyActive;
     bytes32 _conduitKeyCache;
@@ -19,21 +20,28 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
     Consideration private _opensea;
     ConduitController private _conduitController;
 
-    uint lastPartialOrderIndex;
-    uint pendingPartialOrderIndex;
-    mapping(uint => AdvancedOrder) pendingPartialOrders;
-    mapping(uint => uint) pendingPartialOrdersAmount;
+    uint256 lastPartialOrderIndex;
+    uint256 pendingPartialOrderIndex;
+    mapping(uint256 => AdvancedOrder) pendingPartialOrders;
+    mapping(uint256 => uint256) pendingPartialOrdersAmount;
 
-    uint pendingCancelationIndex;
-    uint lastCanceledIndex;
-    mapping(uint => OrderComponents) pendingCancelation;
+    uint256 pendingCancelationIndex;
+    uint256 lastCanceledIndex;
+    mapping(uint256 => OrderComponents) pendingCancelation;
 
     constructor() EchidnaUtils() {
-        _conduitController =  new ConduitController();
+        _conduitController = new ConduitController();
         _opensea = new Consideration(address(_conduitController));
         _conduitKeyCache = bytes32(uint256(uint160(address(this))) << 96);
-        _conduit = _conduitController.createConduit(_conduitKeyCache, address(this));
-        _conduitController.updateChannel(address(_conduit), address(_opensea), true);
+        _conduit = _conduitController.createConduit(
+            _conduitKeyCache,
+            address(this)
+        );
+        _conduitController.updateChannel(
+            address(_conduit),
+            address(_opensea),
+            true
+        );
         require(_conduit == deriveConduit(_conduitKeyCache));
     }
 
@@ -56,7 +64,8 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
             bytes32 orderHash = _opensea.getOrderHash(orderComponent[0]);
             assert(orderHash != bytes32(0));
             // Canceled orders should not be valid
-            (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea.getOrderStatus(orderHash);
+            (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea
+                .getOrderStatus(orderHash);
             assert(!valid);
             assert(canceled);
             assert(filled == size);
@@ -65,19 +74,30 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
 
     function testPartialOrders() public {
         if (lastPartialOrderIndex < pendingPartialOrderIndex) {
-            AdvancedOrder memory order = pendingPartialOrders[lastPartialOrderIndex];
-            uint remaining = pendingPartialOrdersAmount[lastPartialOrderIndex++];
+            AdvancedOrder memory order = pendingPartialOrders[
+                lastPartialOrderIndex
+            ];
+            uint256 remaining = pendingPartialOrdersAmount[
+                lastPartialOrderIndex++
+            ];
             order.numerator = 1;
             order.denominator = 1;
             CriteriaResolver[] memory resolvers = new CriteriaResolver[](0);
-            bool res = _opensea.fulfillAdvancedOrder{value: address(this).balance}(order, resolvers, _conduitKeyActive);
+            bool res = _opensea.fulfillAdvancedOrder{
+                value: address(this).balance
+            }(order, resolvers, _conduitKeyActive);
             assert(res);
             uint256 nonce = _opensea.getNonce(_seller);
-            OrderComponents memory orderComponent = convertOrderParametersToOrderComponents(order.parameters, nonce);
+            OrderComponents
+                memory orderComponent = convertOrderParametersToOrderComponents(
+                    order.parameters,
+                    nonce
+                );
             bytes32 orderHash = _opensea.getOrderHash(orderComponent);
             assert(orderHash != bytes32(0));
             // Filled orders should be valid, not cancelled, and entirely filled.
-            (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea.getOrderStatus(orderHash);
+            (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea
+                .getOrderStatus(orderHash);
             assert(valid);
             assert(!canceled);
             assert(filled != 0);
@@ -85,18 +105,22 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
         }
     }
 
-    function testFulfillAdvancedOrder(bytes32 seed, uint120 numerator, uint120 denominator) public payable {
+    function testFulfillAdvancedOrder(
+        bytes32 seed,
+        uint120 numerator,
+        uint120 denominator
+    ) public payable {
         // FULL_OPEN: 0, PARTIAL_OPEN: 1
-        uint orderType = uint(seed) % 2; 
+        uint256 orderType = uint256(seed) % 2;
 
         // For partial orders we validate the fraction
         // to avoid BadFraction and InexactFraction reverts
         if (orderType == 1) {
-            uint amount = uint256(uint112(uint256(seed)));
-            require(numerator < denominator && numerator != 0); 
+            uint256 amount = uint256(uint112(uint256(seed)));
+            require(numerator < denominator && numerator != 0);
             uint256 valueTimesNumerator = amount * numerator;
             bool exact;
-            uint newValue;
+            uint256 newValue;
             assembly {
                 newValue := div(valueTimesNumerator, denominator)
                 exact := iszero(mulmod(amount, numerator, denominator))
@@ -104,21 +128,29 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
             require(exact);
         }
         // Evenly distribute route between 0 and 5
-        uint256 route = uint(seed) % (6); 
-        (OrderParameters memory orderParams, uint totalTokens, uint totalItems, uint uniqueId) = createOrderParameters(_seller, _buyer, seed, route, false);
+        uint256 route = uint256(seed) % (6);
+        (
+            OrderParameters memory orderParams,
+            uint256 totalTokens,
+            uint256 totalItems,
+            uint256 uniqueId
+        ) = createOrderParameters(_seller, _buyer, seed, route, false);
         orderParams.conduitKey = _conduitKeyActive;
 
         // Sign order on behalf of seller
         uint256 nonce = _opensea.getNonce(_seller);
-        bytes32 orderHash = _opensea.getOrderHash(convertOrderParametersToOrderComponents(orderParams, nonce));
+        bytes32 orderHash = _opensea.getOrderHash(
+            convertOrderParametersToOrderComponents(orderParams, nonce)
+        );
         (, bytes32 domainSeparator, ) = _opensea.information();
         bytes memory sig = signOrder(orderHash, domainSeparator);
-        
+
         // Send entire balance for ether orders (should refund)
-        uint offerItemType = uint(orderParams.offer[0].itemType);
-        uint value = offerItemType < 2 ? address(this).balance : 0;
+        uint256 offerItemType = uint256(orderParams.offer[0].itemType);
+        uint256 value = offerItemType < 2 ? address(this).balance : 0;
         AdvancedOrder memory order;
-        if (orderType == 0) /*FULL_OPEN*/ {
+        if (orderType == 0) /*FULL_OPEN*/
+        {
             order = AdvancedOrder({
                 parameters: orderParams,
                 signature: sig,
@@ -126,7 +158,9 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
                 denominator: uint120(1),
                 extraData: abi.encode(bytes32(0))
             });
-        } else /*PARTIAL_OPEN*/ {
+        }
+        /*PARTIAL_OPEN*/
+        else {
             order = AdvancedOrder({
                 parameters: orderParams,
                 signature: sig,
@@ -135,74 +169,119 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
                 extraData: abi.encode(bytes32(0))
             });
             // Scale order to fill fractional amount
-            uint remaining = totalTokens - ((totalTokens * numerator) / denominator);
+            uint256 remaining = totalTokens -
+                ((totalTokens * numerator) / denominator);
             totalTokens -= remaining;
             pendingPartialOrders[pendingPartialOrderIndex] = order;
             pendingPartialOrdersAmount[pendingPartialOrderIndex++] = remaining;
         }
         // This has no effect without providing a merkle root
-        CriteriaResolver[] memory resolvers = new CriteriaResolver[](0); 
-        try _opensea.fulfillAdvancedOrder{value: value}(order, resolvers, _conduitKeyActive) returns(bool res) {
+        CriteriaResolver[] memory resolvers = new CriteriaResolver[](0);
+        try
+            _opensea.fulfillAdvancedOrder{ value: value }(
+                order,
+                resolvers,
+                _conduitKeyActive
+            )
+        returns (bool res) {
             assert(res);
-        }
-        catch Panic(uint reason) {
+        } catch Panic(uint256 reason) {
             emitAndFail("_opensea.fulfillAdvancedOrder FAILED", route, reason);
         }
 
         // Check that buyers and sellers received expected amounts
-        _assertFundsReceived(_seller, _buyer, route, totalTokens, totalItems, uniqueId);
+        _assertFundsReceived(
+            _seller,
+            _buyer,
+            route,
+            totalTokens,
+            totalItems,
+            uniqueId
+        );
     }
 
     function testFulfillBasicOrder(bytes32 seed) public payable {
         uint256 route = uint256(seed) % 6;
         uint256 orderType = uint256(seed) % 2;
-         BasicOrderType basicOrderType = BasicOrderType(
-            orderType + (4 * route)
-        );
-        (OrderParameters memory orderParams, uint totalTokens, uint totalItems, uint uniqueId) = createOrderParameters(_seller, _buyer, seed, route, true);
+        BasicOrderType basicOrderType = BasicOrderType(orderType + (4 * route));
+        (
+            OrderParameters memory orderParams,
+            uint256 totalTokens,
+            uint256 totalItems,
+            uint256 uniqueId
+        ) = createOrderParameters(_seller, _buyer, seed, route, true);
         orderParams.conduitKey = _conduitKeyActive;
 
         // Sign order on behalf of seller
         uint256 nonce = _opensea.getNonce(_seller);
-        bytes32 orderHash = _opensea.getOrderHash(convertOrderParametersToOrderComponents(orderParams, nonce));
+        bytes32 orderHash = _opensea.getOrderHash(
+            convertOrderParametersToOrderComponents(orderParams, nonce)
+        );
         (, bytes32 domainSeparator, ) = _opensea.information();
         bytes memory sig = signOrder(orderHash, domainSeparator);
-        
-        BasicOrderParameters memory basicOrder = convertOrderParametersToBasicOrder(orderParams, sig, basicOrderType);
+
+        BasicOrderParameters
+            memory basicOrder = convertOrderParametersToBasicOrder(
+                orderParams,
+                sig,
+                basicOrderType
+            );
         // Send entire balance for ether orders (should refund)
         uint256 value = uint256(basicOrderType) < 8 ? address(this).balance : 0;
-        (bool success, ) = address(_opensea).call{ value: value }(abi.encodeWithSelector(Consideration.fulfillBasicOrder.selector, basicOrder));
+        (bool success, ) = address(_opensea).call{ value: value }(
+            abi.encodeWithSelector(
+                Consideration.fulfillBasicOrder.selector,
+                basicOrder
+            )
+        );
         if (!success) {
             emitAndFail("fulfillBasicOrder", orderParams.offer.length, route);
         }
         // totalItems will always be 1 for basic orders
         // and used for ERC721 balances only
         // otherwise InvalidERC721TransferAmount revert
-        _assertFundsReceived(_seller, _buyer, route, totalTokens, totalItems, uniqueId);
+        _assertFundsReceived(
+            _seller,
+            _buyer,
+            route,
+            totalTokens,
+            totalItems,
+            uniqueId
+        );
     }
-
 
     function testValidate(bytes32 seed) public {
         Order[] memory orders = new Order[](1);
-        uint route =  uint(seed) % 6;
-        (OrderParameters memory param,,,) = createOrderParameters(payable(address(this)), payable(address(this)), seed, route, false);
+        uint256 route = uint256(seed) % 6;
+        (OrderParameters memory param, , , ) = createOrderParameters(
+            payable(address(this)),
+            payable(address(this)),
+            seed,
+            route,
+            false
+        );
         // Don't sign these because to cancel them
         // the offerer must be msg.sender
         orders[0] = Order({
             parameters: param,
-            signature: abi.encode(bytes32(0)) 
+            signature: abi.encode(bytes32(0))
         });
 
         bool res = _opensea.validate(orders);
-        assert(res);    
+        assert(res);
 
         uint256 nonce = _opensea.getNonce(address(this));
-        OrderComponents memory orderComponent = convertOrderParametersToOrderComponents(orders[0].parameters, nonce);
+        OrderComponents
+            memory orderComponent = convertOrderParametersToOrderComponents(
+                orders[0].parameters,
+                nonce
+            );
         bytes32 orderHash = _opensea.getOrderHash(orderComponent);
         assert(orderHash != bytes32(0));
 
         // Validated orders should be fillable and not canceled
-        (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea.getOrderStatus(orderHash);
+        (bool valid, bool canceled, uint256 filled, uint256 size) = _opensea
+            .getOrderStatus(orderHash);
         assert(valid);
         assert(!canceled);
         assert(filled == size);
@@ -212,23 +291,30 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
     }
 
     function testIncrementNonce() public {
-        // Retrieve the old nonce 
+        // Retrieve the old nonce
         uint256 oldNonce = _opensea.getNonce(address(this));
-        // Try incrementing it 
+        // Try incrementing it
         (bool success, ) = address(_opensea).call(
             abi.encodeWithSelector(Consideration.incrementNonce.selector)
-        );        
+        );
         if (!success) {
-            emitAndFail("_opensea.incrementNonce() call failed", oldNonce, oldNonce);
+            emitAndFail(
+                "_opensea.incrementNonce() call failed",
+                oldNonce,
+                oldNonce
+            );
         }
         // Retrieve the updated nonce
         uint256 newNonce = _opensea.getNonce(address(this));
-        // oldNonce should never exceed or equal newNonce 
-        if (oldNonce > newNonce) { 
-            emitAndFail("oldNonce > newNonce after incrementing", oldNonce, newNonce);
-        }           
-        else if (oldNonce == newNonce) {
-            emitAndFail("oldNonce is equal to newNonce", oldNonce, newNonce);            
+        // oldNonce should never exceed or equal newNonce
+        if (oldNonce > newNonce) {
+            emitAndFail(
+                "oldNonce > newNonce after incrementing",
+                oldNonce,
+                newNonce
+            );
+        } else if (oldNonce == newNonce) {
+            emitAndFail("oldNonce is equal to newNonce", oldNonce, newNonce);
         }
     }
 
@@ -238,7 +324,7 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
         uint256 id,
         uint256 amount,
         bytes calldata data
-    ) external override returns(bytes4) {
+    ) external override returns (bytes4) {
         return ERC1155TokenReceiver.onERC1155Received.selector;
     }
 
@@ -252,7 +338,10 @@ contract Echidna is ERC1155TokenReceiver, EchidnaUtils {
         return ERC1155TokenReceiver.onERC1155BatchReceived.selector;
     }
 
-    function deriveConduit(bytes32 conduitKey) internal returns(address conduit) {
+    function deriveConduit(bytes32 conduitKey)
+        internal
+        returns (address conduit)
+    {
         bytes32 conduitCreationCodeHash = keccak256(type(Conduit).creationCode);
         address conduitController = address(_conduitController);
         assembly {
