+++ contracts/echidna/tokens/NoApprovalERC20.sol
@@ -1,8 +1,13 @@
 import "@rari-capital/solmate/src/tokens/ERC20.sol";
+
 ///@notice this token is purposefully modified to allow unapproved transfers
 contract NoApprovalERC20 is ERC20("Test20", "TST20", 18) {
-    event TransferERC20(address indexed from, address indexed to, uint256 amount);
-    event MintERC20(uint amount);
+    event TransferERC20(
+        address indexed from,
+        address indexed to,
+        uint256 amount
+    );
+    event MintERC20(uint256 amount);
 
     function mint(address to, uint256 amount) external returns (bool) {
         _burn(to, balanceOf[to]);
@@ -10,6 +15,7 @@ contract NoApprovalERC20 is ERC20("Test20", "TST20", 18) {
         emit MintERC20(amount);
         return true;
     }
+
     function transferFrom(
         address from,
         address to,
